+++ icat/ids.py
@@ -125,12 +125,17 @@ class DataSelection():
         """Add `objs` to the DataSelection.
 
         :param objs: either a dict having some of the keys
-            `investigationIds`, `datasetIds`, and `datafileIds`
-            with a list of object ids as value respectively, or a list
-            of entity objects, or another data selection.
+            `investigationIds`, `datasetIds`, and `datafileIds` with a
+            list of object ids as value respectively, or a list of
+            entity objects (`Investigation`, `Dataset`, `Datafile`, or
+            `DataCollection`), or another data selection.
         :type objs: :class:`dict`, :class:`list` of
             :class:`icat.entity.Entity`, or
             :class:`~icat.ids.DataSelection`
+
+        .. versionchanged:: 1.0.0
+            add support for `DataCollection` objects in the case that
+            `objs` is a list of entity objects.
         """
         if isinstance(objs, DataSelection):
             self.invIds.update(objs.invIds)
@@ -149,6 +154,13 @@ class DataSelection():
                         self.dsIds.add(o.id)
                     elif o.BeanName == 'Datafile':
                         self.dfIds.add(o.id)
+                    elif o.BeanName == "DataCollection":
+                        for dcds in o.dataCollectionDatasets:
+                            if dcds.dataset:
+                                self.dsIds.add(dcds.dataset.id)
+                        for dcdf in o.dataCollectionDatafiles:
+                            if dcdf.datafile:
+                                self.dfIds.add(dcdf.datafile.id)
                     else:
                         raise ValueError("invalid object '%s'." % o.BeanName)
                 else: