+++ CHANGES.rst
@@ -5,6 +5,12 @@ Changelog
 1.0.0 (not yet released)
 ~~~~~~~~~~~~~~~~~~~~~~~~
 
+New features
+------------
+
++ `#74`_: :class:`icat.ids.DataSelection` also accepts
+  `DataCollection` as argument.
+
 Incompatible changes and deprecations
 -------------------------------------
 
@@ -32,6 +38,8 @@ Bug fixes and minor changes
 ---------------------------
 
 + Some (more) example scripts now require ICAT 4.4.0 or newer.
+
+.. _#74: https://github.com/icatproject/python-icat/issues/74
 
 
 0.17.0 (2020-04-30)